Reorganize api files to respective folders
