--- api/gifs.json
@@ -1,262 +1,0 @@
-[
-  {
-    "name": "Audrey Hepburn",
-    "img": "audrey-hepburn-nounish-what.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "B8617CE1-D5F0-4C7B-AF50-6A9111808A1C.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "blinking_v4.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "breakfast-club-smoke-nounish.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "C71C7305-849F-478F-BEBB-CA47F9932484.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "carl-simpsons-nounish-chefskiss_1.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "cdt_coffee_01.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "chris-farley-nounish_v2.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "crytears_nounish2.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "Eric-Andre-NOun-Rules.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "hank-hill-hose4.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "hank-hill-screaming-nounish3.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "homer-hedge-nounish.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "jaw-drop-shocked-1-2.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "jay-z-nounish-headbob.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "jay-z-nounpunkish-headbob.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "jurassic_park_take_off_nouns_glasses_1.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "jurassic_park_take_off_nouns_glasses_infinite_1.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "jurassic_park_take_off_nouns_glasses_loop_3.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "jurassic_park_take_off_nouns_glasses_reverse_1.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "jurassic_park_take_off_nouns_sunglasses_1.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "ken-jeong-small_nounish.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "king-of-the-hill-jpeg-nounish1.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "lenny-thinking-nounish2.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "matrix-nounish-neo-s.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "MIB_I-make-this-look-good-2.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "milhouse-simpsons-eyebrows-nounish2.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "nicholsnoun_classic.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "Noun-Life-Magazine-Blink1.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "nounblinking460.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "noungm_2.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "nounish_bud_1.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "nounish_bud_light_1.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "nounish_suspicious_fry2.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "nounish-building-wallace-gromit-3.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "nounish-pokemon-squirtle2.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "nounish-suspicious-monkey-puppet.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "nounish-toad-frog-bud-budweiser-1.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "nounish-toad-scared-excited.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "Nouns_Brave_Empire_blink.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "Nyoun_Cat.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "noun-thousand2.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "noun-thousand.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "redford.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "robert-redford-nod-nounish-4.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "seriously1.png",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "shocked-nounish-david-rose-2.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "shutup-and-take-my-money.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "spider-man-you.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "spongebob-lonely-1-nounish.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "up-goggles-3csm.gif",
-    "type": "gif"
-  },
-  {
-    "name": "",
-    "img": "woah-removing-noun4.gif",
-    "type": "gif"
-  }
-]
