--- api/nav.json
@@ -1,58 +1,0 @@
-[
-  {
-    "name": "Intro",
-    "current": false,
-    "link": "/intro"
-  },
-  {
-    "name": "Projects",
-    "current": false,
-    "link": "/projects"
-  },
-  {
-    "name": "Dev",
-    "current": false,
-    "link": "/dev"
-  },
-  {
-    "name": "Nouners",
-    "current": false,
-    "link": "/nouners"
-  },
-  {
-    "name": "History",
-    "current": false,
-    "link": "/history"
-  },
-  {
-    "name": "Funding",
-    "current": false,
-    "link": "/funding"
-  },
-  {
-    "name": "Traits",
-    "current": false,
-    "link": "/traits"
-  },
-  {
-    "name": "Podcast",
-    "current": false,
-    "link": "/podcast"
-  },
-  {
-    "name": "Talent",
-    "current": false,
-    "link": "/talent"
-  },
-
-  {
-    "name": "Notes",
-    "current": false,
-    "link": "/notes"
-  },
-  {
-    "name": "Assets",
-    "current": false,
-    "link": "/assets"
-  }
-]
