+++ components/Game/DinoScript.js
