+++ components/Game/DinoStyle.js
