+++ components/Layout.tsx
@@ -1,9 +1,9 @@
-import NewNav from './NewNav/NewNav';
+import Navigation from './Navigation/Web';
 
 export default function Layout({ children }) {
   return (
     <>
-      <NewNav />
+      <Navigation />
 
       <div className=" flex flex-col flex-1 flex-between h-auto text-black m-auto">
         <main className="w-full  max-w-5xl	m-auto DocSearch-content">
