+++ components/Navigation/navigation.json
@@ -21,7 +21,6 @@
         "link": "/history",
         "external": false
       },
-   
       {
         "name": "Discord Channels",
         "link": "/channels",
