+++ pages/assets/index.tsx
@@ -1,13 +1,13 @@
 /* eslint-disable @next/next/no-img-element */
 import React from 'react';
-import Header from '../components/Header';
-import Title from '../components/Title';
+import Header from '../../components/Header';
+import Title from '../../components/Title';
 import { v4 as uuidv4 } from 'uuid';
-import assets from '../api/assets.json';
-import Subheader from '../components/Subheader';
-import PageHeader from '../components/Layout/PageHeader';
-import PageContent from '../components/Layout/PageContent';
-import Button from '../components/common/Button';
+import assets from './assets.json';
+import Subheader from '../../components/Subheader';
+import PageHeader from '../../components/Layout/PageHeader';
+import PageContent from '../../components/Layout/PageContent';
+import Button from '../../components/common/Button';
 
 const Assets = () => {
   return (
