+++ pages/channels/index.tsx
@@ -1,11 +1,11 @@
 /* eslint-disable @next/next/no-img-element */
 import React from 'react';
-import Header from '../components/Header';
-import Title from '../components/Title';
-import Subheader from '../components/Subheader';
-import PageHeader from '../components/Layout/PageHeader';
-import PageContent from '../components/Layout/PageContent';
-import channelGroups from '../api/channels.json';
+import Header from '../../components/Header';
+import Title from '../../components/Title';
+import Subheader from '../../components/Subheader';
+import PageHeader from '../../components/Layout/PageHeader';
+import PageContent from '../../components/Layout/PageContent';
+import channelGroups from './channels.json';
 
 const Channels = () => {
   return (
