+++ pages/dev/resources.json
