+++ pages/talent/talent.json
@@ -210,7 +210,7 @@
   {
     "name": "Rye",
     "twitter": "ryeshrimp",
-    "discord": "ryeshrimp#8056",
+    "discord": "ryeshrimp#0001",
     "discordId": "388565991065649152",
     "title": ["developer"],
     "skills": "Full stack, Solidity, Design, UX, Project Management"
